30.01.25

+++ lib/v2/task/CT.php
@@ -6,7 +6,7 @@ use Iplogic\Beru\V2\Task;
 use Iplogic\Beru\V2\ORM\TaskTable;
 
 /**
- * Checks for old unfinished tasks and deletes them
+ * Checks for old and unfinished tasks and deletes them
  *
  * Class CT
  * @package Iplogic\Beru\V2\Task
@@ -16,12 +16,19 @@ class CT implements TaskInterface
 
 	public function execute($arTask): void
 	{
-		$time = time() - 300;
+		$time = time() - 3000;
 		$result = TaskTable::getList(["filter" => ["<=UNIX_TIMESTAMP" => $time, "=STATE" => "IW"]]);
 		while( $task = $result->fetch() ) {
 			TaskTable::delete($task["ID"]);
 		}
-		TaskTable::delete($arTask["ID"]);
+		$time = time() - 2592000;
+		$result = TaskTable::getList(["filter" => ["<=UNIX_TIMESTAMP" => $time]]);
+		while( $task = $result->fetch() ) {
+			TaskTable::delete($task["ID"]);
+		}
+		if(isset($arTask["ID"]) && $arTask["ID"] > 0) {
+			TaskTable::delete($arTask["ID"]);
+		}
 		Task::scheduleTask(0, "CT", 600);
 		return;
 	}
